Move routes dir
